Data.X509.Extra

Description

Cryptographic & Data.X509 specialized methods for RSA with SHA256

RSA/SHA-256 Applied Constructors

signAlgRSA256 ∷ SignatureALG

Algorithm Signature for RSA with SHA256

signCertificate ∷ (MonadFail m, MonadRandom m) ⇒ PrivateKey → Certificate → m SignedCertificate

Sign a X.509 certificate using RSA-PKCS1.5 with SHA256

genRSA256KeyPair ∷ IO (PublicKey, PrivateKey)

Generate a new RSA-256 key pair

validateDefaultWithIP ∷ CertificateStore → ValidationCache → ServiceID → CertificateChain → IO [FailedReason]

Drop-in replacement for validateDefault but with support for IP SAN

validateCertificate ∷ SignedCertificate → ValidationChecks → ServiceID → SignedCertificate → IO [FailedReason]

Validate a X.509 certificate using SHA256 hash and a given CA. This is merely to verify that we aren't generating invalid certificates.

Utils

failIfReasons ∷ MonadFail m ⇒ [FailedReason] → m ()

Fail with the given reason if any, does nothing otherwise

parseSAN ∷ String → AltName

Parse a Subject Alternative Name (SAN) from a raw string

isServerCertificate ∷ SignedCertificate → Bool

Tell whether a Certificate is a Server certificate

isClientCertificate ∷ SignedCertificate → Bool

Tell whether a Certificate is Client certificate

RSA Encode PEM

class EncodePEM a where

Encode an artifact to PEM (i.e. base64 DER with header & footer)

Minimal complete definition

encodePEM

Methods

encodePEM ∷ a → ByteString

encodePEMRaw ∷ (ByteString, a → ByteString, ByteString) → a → ByteString

Effectful IO Functions

writeCredentials ∷ FilePath → (PrivateKey, SignedCertificate) → IO ()

Write a certificate and its private key to the given location

writeCertificate ∷ FilePath → SignedCertificate → IO ()

Write a certificate to the given location

Re-Export