Main

Documentation

main ∷ IO ()

propGenCertificateValid ∷ (TLSConfiguration, DirConfiguration) → Property

Verify that each certificate generated is valid. Is uses the default validation check of Validation

propUnknownService ∷ Unknown AltNames → (TLSConfiguration, DirConfiguration) → Property

Verify that each server certificate generated is invalid when provided an unknown ServiceID.

propInvalidExpiryDays ∷ (Invalid TLSConfiguration, DirConfiguration) → Property

Verify that we can't generate certificates when provided invalid expiry days.

propAllCertsValid ∷ SignedCertificate → [FailedReason] → Property

Check that there's no validation FailedReason

propServerCertsInvalid ∷ SignedCertificate → [FailedReason] → Property

Check that there are actually some validation FailedReason for non-client certificate

generateAndValidate ∷ (TLSConfiguration → ServiceID) → (SignedCertificate → [FailedReason] → Property) → (TLSConfiguration, DirConfiguration) → IO Property

Actually generate certificates and validate them with the given property Throws on error.

getValidServiceID ∷ TLSConfiguration → ServiceID

Get a valid serviceID from the configuration

getUnknownServiceID ∷ Unknown AltNames → TLSConfiguration → ServiceID

Get an invalid serviceID from the configuration

(=/=) ∷ (Eq a, Show a) ⇒ a → a → Property infix 4

Like /=, but prints a counterexample when it fails. Source: QuickCheck@2.12 Test.QuickCheck.Property#(=/=)