Test.Cardano.X509.Configuration.Arbitrary

Documentation

newtype Show a ⇒ Invalid a

Helper to declare Arbitrary instances with generators that generate invalid values

Constructors

Invalid

Fields

getInvalid ∷ a

newtype Show a ⇒ Unknown a

Helper to declare Arbitrary instances with generators that generate unkown valus

Constructors

Unknown

Fields

getUnknown ∷ a

newtype AltNames

Easily generate alternative names

Constructors

AltNames

Fields

getAltNames ∷ NonEmpty String

Orphan instances

Arbitrary TLSConfiguration

Instance details

Methods

arbitrary ∷ Gen TLSConfiguration

shrink ∷ TLSConfiguration → [TLSConfiguration]

Arbitrary DirConfiguration

Instance details

Methods

arbitrary ∷ Gen DirConfiguration

shrink ∷ DirConfiguration → [DirConfiguration]

Arbitrary CertConfiguration

Instance details

Methods

arbitrary ∷ Gen CertConfiguration

shrink ∷ CertConfiguration → [CertConfiguration]

Arbitrary ServerConfiguration

Instance details

Methods

arbitrary ∷ Gen ServerConfiguration

shrink ∷ ServerConfiguration → [ServerConfiguration]